requirements.txt
docopt
pykwalify
pyyaml
pytest